+++ modules/hanna-react/src/utils.ts
@@ -1,2 +1,3 @@
 export * from './utils/useFormatMonitor';
 export * from './utils/useGetSVGtext';
+export * from './utils/useScrollbarWidthCSSVar';