+++ modules/hanna-react/CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Upcoming...
 
 - ... <!-- Add new lines here. -->
+- feat: Add `utils` hook `useDidChange`
 - feat: Add `utils` hook `useScrollbarWidthCSSVar`
 - fix: Stop hard-resetting `AccordionList`'s state on `defaultOpen` changes